turning library into its own dynamic/client componetns
