+++ app/dashboard/library/get-library-images.tsx
@@ -1,0 +1,8 @@
+import { getImagesAction } from '@/actions/dashboard-actions'
+import LibraryClient from './library-client'
+
+export async function GetLibraryImages() {
+  const images = await getImagesAction()
+
+  return <LibraryClient images={images} />
+}
